RavenDB-18204 : address PR comments
